fix(#119): modules moving of place after startup

+++ scripts/lefthook/pre_hook.rs
@@ -4,8 +4,10 @@ use std::fs;
 fn main() {
     let dist = std::path::PathBuf::from("dist");
     let temp = std::path::PathBuf::from(".dist");
-    if dist.exists() {
+    if dist.exists() && !temp.exists() {
         fs::rename(&dist, temp).unwrap();
     }
-    fs::create_dir(dist).unwrap();
+    if !dist.exists() {
+        fs::create_dir(dist).unwrap();
+    }
 }
